refactor: server.ts와 client.ts에 `util`을 앞에 붙여주기
