+++ packages/web/src/app/[lng]/(auth)/memos/_components/MemoSearchForm/index.tsx
@@ -3,7 +3,7 @@
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@extension/ui';
 import { Input } from '@src/components/ui';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { X } from 'lucide-react';
 import { Controller, useFormContext } from 'react-hook-form';
 import { SearchFormValues } from '../MemoSearchFormProvider';
