+++ packages/web/src/app/[lng]/(auth)/memos/_components/MemoSidebar/SidebarGroupCategory.tsx
@@ -11,7 +11,7 @@ import {
   SidebarMenuItem,
 } from '@src/components/ui';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { useRouter } from 'next/navigation';
 import { memo } from 'react';
