+++ packages/web/src/app/[lng]/(auth)/memos/_components/MemoSidebarTrigger/index.tsx
@@ -2,7 +2,7 @@
 
 import { SidebarTrigger, Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@src/components/ui';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/server';
+import useTranslation from '@src/modules/i18n/util.server';
 
 interface MemoSidebarTriggerProps extends LanguageType {}
