+++ packages/web/src/app/[lng]/(auth)/memos/_components/MemoView/MemoRefreshButton.tsx
@@ -4,7 +4,7 @@ import { setLocalStorageTrue } from '@extension/shared/modules/local-storage';
 import { Button, toast } from '@src/components/ui';
 import { useGuide } from '@src/modules/guide';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { useQueryClient } from '@tanstack/react-query';
 import { RefreshCwIcon } from 'lucide-react';
 import { memo } from 'react';
