+++ packages/web/src/app/[lng]/(auth)/memos/_components/MemoView/ToggleView.tsx
@@ -2,7 +2,7 @@
 import { useSearchParams } from '@extension/shared/modules/search-params';
 import { ToggleGroup, ToggleGroupItem } from '@src/components/ui';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { Calendar, Grid } from 'lucide-react';
 import { useRouter } from 'next/navigation';
 import { memo } from 'react';
