+++ packages/web/src/app/[lng]/(auth)/memos/setting/_components/Setting/SettingCategoryForm.tsx
@@ -6,7 +6,7 @@ import {
 } from '@extension/shared/hooks';
 import { Button, Input, Label, toast } from '@src/components/ui';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { TrashIcon } from 'lucide-react';
 import { useEffect } from 'react';
 import { useForm } from 'react-hook-form';
