+++ packages/web/src/app/[lng]/(auth)/memos/setting/_components/Setting/SettingGuide.tsx
@@ -2,7 +2,7 @@ import { PATHS } from '@extension/shared/constants';
 import { LocalStorage } from '@extension/shared/modules/local-storage';
 import { Button, Label } from '@src/components/ui';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { useRouter } from 'next/navigation';
 
 interface SettingGuideProps extends LanguageType {}
