+++ packages/web/src/app/[lng]/(auth)/memos/setting/_components/Setting/SettingLanguage.tsx
@@ -1,7 +1,7 @@
 import { Label, Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@src/components/ui';
 import { useLanguage } from '@src/hooks';
 import { Language, LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { useRouter } from 'next/navigation';
 
 interface SettingLanguageProps extends LanguageType {}
