+++ packages/web/src/app/[lng]/(auth)/memos/setting/_components/SettingHeader/index.tsx
@@ -2,7 +2,7 @@
 
 import { Label } from '@src/components/ui';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/server';
+import useTranslation from '@src/modules/i18n/util.server';
 
 interface SettingHeaderProps extends LanguageType {}
