+++ packages/web/src/app/[lng]/(no-auth)/introduce/_components/AdditionalFeatures/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { motion } from 'framer-motion';
 
 interface AdditionalFeaturesProps extends LanguageType {}
