+++ packages/web/src/app/[lng]/(no-auth)/introduce/_components/QuestionAndAnswer/index.tsx
@@ -2,7 +2,7 @@
 
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@extension/ui';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { motion } from 'framer-motion';
 import { FAQ_ITEMS } from '../../_constants/Faq';
