+++ packages/web/src/app/[lng]/(no-auth)/login/_components/LoginSection/index.tsx
@@ -3,7 +3,7 @@
 import { SUPABASE } from '@extension/shared/constants';
 import { Button } from '@src/components/ui';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/server';
+import useTranslation from '@src/modules/i18n/util.server';
 import { signInWithEmail, signInWithOAuth } from '@src/modules/supabase/util.server';
 import Image from 'next/image';
