+++ packages/web/src/app/[lng]/(no-auth)/update/_components/UpdateTitle/index.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
+import useTranslation from '@src/modules/i18n/util.client';
 import { motion } from 'framer-motion';
 
 interface UpdateTitleProps extends LanguageType {}
