+++ packages/web/src/app/[lng]/_components/index.ts
@@ -1,3 +1,2 @@
-export { default as Header } from '../../../components/Header';
 export { default as ThemeProvider } from './ThemeProvider';
 export { default as QueryProvider } from './QueryProvider';
