+++ packages/web/src/modules/guide/useGuide.ts
@@ -6,7 +6,7 @@ import { driver } from 'driver.js';
 import { useEffect } from 'react';
 
 import { LanguageType } from '../i18n';
-import useTranslation from '../i18n/client';
+import useTranslation from '../i18n/util.client';
 
 interface UseGuideProps extends LanguageType {}
